backend/requirements.txt
pytesseract
pillow
pymupdf
python-pptx

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2

openai==1.57.0

pdfplumber==0.11.4
python-docx==1.1.2
